Issue-9565 - Port ArrayRepeat to function-arrays subcrate
