+++ datafusion/functions-array/src/lib.rs
@@ -55,6 +55,7 @@ pub mod expr_fn {
     pub use super::udf::array_empty;
     pub use super::udf::array_length;
     pub use super::udf::array_ndims;
+    pub use super::udf::array_repeat;
     pub use super::udf::array_sort;
     pub use super::udf::array_to_string;
     pub use super::udf::cardinality;
@@ -86,6 +87,7 @@ pub fn register_all(registry: &mut dyn FunctionRegistry) -> Result<()> {
         udf::flatten_udf(),
         udf::array_sort_udf(),
         udf::array_distinct_udf(),
+        udf::array_repeat_udf(),
     ];
     functions.into_iter().try_for_each(|udf| {
         let existing_udf = registry.register_udf(udf)?;
