+++ datafusion/proto/tests/cases/roundtrip_logical_plan.rs
@@ -558,6 +558,7 @@ async fn roundtrip_expr_api() -> Result<()> {
         ),
         array_empty(make_array(vec![lit(1), lit(2), lit(3)])),
         array_length(make_array(vec![lit(1), lit(2), lit(3)])),
+        array_repeat(lit(1), lit(3)),
         flatten(make_array(vec![lit(1), lit(2), lit(3)])),
         array_sort(
             make_array(vec![lit(3), lit(4), lit(1), lit(2)]),
